refactor)folder-structure): migrating to individual _components folders
