+++ app/_components/Navbar/index.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link"
 // COMPONENTS
-import NavBarAvatar from "./NavBarAvatar"
-import UserNavProOption from "./UserNavProOption"
+import { NavbarAvatar } from "./navbar-avatar"
+import { NavbarAuthedOptions } from "./navbar-authed-options";
 
 export default async function Navbar() {
   return (
@@ -16,7 +16,7 @@ export default async function Navbar() {
         </Link>
       </div>
       <ul className="flex-1 flex justify-end items-center gap-10 content-between">
-        <UserNavProOption />
+        <NavbarAuthedOptions />
         <li className="hover:scale-105 hover:opacity-75 font-semibold">
           <Link href={"/blog"}>
             Blog
@@ -27,7 +27,7 @@ export default async function Navbar() {
             Courses
           </Link>
         </li>
-        <NavBarAvatar />
+        <NavbarAvatar />
       </ul>
     </nav>
   )
