+++ app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 // COMPONENTS
-import Navbar from '@/components/Navbar';
+import Navbar from '@/app/_components/Navbar';
 import { AuthContextProvider } from '@/context/AuthContext';
 // UTILS
 import { fontRoboto, fontPoppins } from '@/utils/font';
