+++ components/Buttons/index.tsx
@@ -1,5 +1,5 @@
 import { CheckoutButton } from "./CheckoutButton";
-import { ManageSubscriptionButton } from "./ManageSubscriptionButton";
+import { ManageSubscriptionButton } from "../../app/dashboard/billing/_components/manage-subscription-button";
 import { ToggleAutoPlayButton } from "./ToggleAutoPlayButton";
 
 export {
